requirements.txt
Flask==1.1.1
firebase-admin==3.2.1
python-dotenv==0.11.0